Feat: adjust the exponent for distance forces
The original simulation uses forces that are *distance-independent*.
That is, if the distance between particles is 3 or 30, the force is
the same. I wanted to experiment with forces that depend slightly on
distance.

I also want to experiment with forces that depend much more on
distance (electromagnetism, van der waals, lennard-jones, mie
potential) but that's going to require tweaking a bunch of the
parameters, I think.